Second iteration of build model
In this iteration, I fix the base, so it has less material and can be printed. Also fixes motor support, getting it smaller an with align marks. Fixed the inner base with new drills. Fixed axis of cap, and shortened the axis, too. Created some tests to fit press fits and motor axis - join.